pyproject.toml
[project]
name = "trading-system"
version = "0.2.0"
description = "NautilusTrader-based algorithmic trading system with IBKR and MT5 support."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nautilus_trader>=1.200",
    "pandas>=2.0,<3.0",
    "pyarrow>=14.0",
    "holidays>=0.40",
    "pyyaml>=6.0",
    "aiosqlite>=0.19",
]

[project.optional-dependencies]
ibkr = [
    "nautilus_trader[ibkr]",
]
metatrader = [
    "MetaTrader5>=5.0",
]
polygon = [
    "polygon-api-client>=1.13",
]
charts = [
    "plotly>=5.0",
]
notebooks = [
    "nbformat>=5.9",
    "jupyterlab>=4.0",
    "plotly>=5.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["trader*", "historical_data_services*"]
